[mlir][Linalg] Generalize identity generic op removal to handle trivial broadcasts.
The existing pattern to remove identity generic ops when input is just
copied to the output looks for strict identity of the indexing
maps. This could be generalized to handle another case where the
indexing maps for the operand and result match, except for existence
of unit dimensions in either. In such cases use reshapes to replace
the generic op.
The existing pattern for removing identity generic op is superceded by
this pattern for ops with tensor semantics. The existing pattern is
now only used for the case where the op has buffer semantics.